Refactor GM services and availability management
- Updated GmDetailsModalComponent to use GmDirectoryService instead of GmService for GM display name retrieval.
- Refactored TechStackComponent to replace GmService with GmDirectoryService.
- Commented out the IAvailabilitySlot interface in i-gm-profile.ts for potential future use.
- Modified SCSS for GM availability to improve layout and styling.
- Added AvailabilityService with methods for managing availability slots, including fetching, upserting, and deleting availability.
- Introduced GmDirectoryService for managing GM data retrieval and display names.
- Implemented GmSchedulingService to handle scheduling logic, including availability checks and slot suggestions.
- Created interfaces for availability slots to standardize data structure.
