+++ src/app/auth/core/services/availability/availability.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AvailabilityService } from './availability.service';
+
+describe('AvailabilityService', () => {
+  let service: AvailabilityService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(AvailabilityService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
