+++ src/app/auth/core/services/gm/gm-directory/gm-directory.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { GmDirectoryService } from './gm-directory.service';
+
+describe('GmDirectoryService', () => {
+  let service: GmDirectoryService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(GmDirectoryService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
