+++ src/app/auth/core/services/gm/gm-scheduling/gm-scheduling.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { GmSchedulingService } from './gm-scheduling.service';
+
+describe('GmSchedulingService', () => {
+  let service: GmSchedulingService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(GmSchedulingService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
