+++ src/app/core/interfaces/i-availability-slot.ts
@@ -1,0 +1,10 @@
+export type WorkType = 'gm' | 'reception' | 'external_event';
+
+export interface IAvailabilitySlot {
+  id?: string;
+  userId: string;
+  date: string;
+  fromHour: number;
+  toHour: number;
+  workType: WorkType;
+}
